@@ -1,11 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 31
-    buildToolsVersion "30.0.3"
+    compileSdk 34
+    buildToolsVersion = "34.0.0"
+    namespace = "com.starseed.speechtotext"
     defaultConfig {
         minSdkVersion 24
-        targetSdkVersion 30
+        targetSdkVersion 35
         versionCode 1
         versionName "1.0"
     }
@@ -21,19 +22,20 @@ android {
 
 dependencies {
     api fileTree(include: ['*.jar'], dir: 'libs')
-    testImplementation 'junit:junit:4.12'
-    api 'androidx.appcompat:appcompat:1.4.2'
-    api 'com.google.android.material:material:1.6.1'
-    compile files('libs/classes.jar')
+    testImplementation 'junit:junit:4.13.2'
+    api 'androidx.appcompat:appcompat:1.7.0'
+    api 'com.google.android.material:material:1.12.0'
+    implementation files('libs/classes.jar')
+    //implementation files('libs/unity-classes.jar')
 }
 
 //task to delete the old jar
-task deleteOldJar(type: Delete) {
+tasks.register('deleteOldJar', Delete) {
     delete 'release/SpeechToTextPlugin.jar'
 }
 
 //task to export content as jar
-task exportJar(type: Copy) {
+tasks.register('exportJar', Copy) {
     from('build/intermediates/aar_main_jar/release/')
     into('release/')
     include('classes.jar')
